# Conflicts:
#	.gitignore
#	mvnw
#	src/main/java/com/q7w/examination/Service/ExroomService.java
#	src/main/java/com/q7w/examination/Service/PaperService.java
#	src/main/java/com/q7w/examination/Service/QuestionsService.java
#	src/main/java/com/q7w/examination/Service/impl/ExroomServiceimpl.java
#	src/main/java/com/q7w/examination/Service/impl/PaperServiceimpl.java
#	src/main/java/com/q7w/examination/Service/impl/QuestionsServiceimpl.java
#	src/main/java/com/q7w/examination/controller/CouseController.java
#	src/main/java/com/q7w/examination/controller/ExroomController.java
#	src/main/java/com/q7w/examination/controller/LoginController.java
#	src/main/java/com/q7w/examination/controller/PaperController.java
#	src/main/java/com/q7w/examination/controller/QuestionsController.java
#	src/main/java/com/q7w/examination/controller/TestController.java
#	src/main/java/com/q7w/examination/dao/PaperDAO.java
#	src/main/java/com/q7w/examination/dto/PaperDTO.java
#	src/main/java/com/q7w/examination/entity/Paper.java
#	src/main/java/com/q7w/examination/entity/Questions.java

+++ src/main/java/com/q7w/examination/entity/Exroom.java
@@ -31,6 +31,6 @@ public class Exroom extends BaseEntity implements Serializable {
     private long deadline;//截至进入时间
     private long time;//考试时间
     private int type;//考场类型
-    private long allowtimes;//考试时长
+    private int allowtimes;//考试时长
 
 }
